% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorepoint.R
\name{add.restore.point.test}
\alias{add.restore.point.test}
\title{Add one or several test functions}
\usage{
add.restore.point.test(...)
}
\arguments{
\item{...}{a slist of test functions that will be called with the stored arguments}
}
\description{
A test function is called after a restore point
has stored data. It must have an argument env and name.
It can check whether certain conditions are satisfied by the variables
}

