% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/restorepoint.R
\name{assert}
\alias{assert}
\title{Checks whether cond holds true if not throws an error}
\usage{
assert(cond)
}
\arguments{
\item{cond}{a condition that is checked}
}
\description{
Can be used for checking for errors in functions
}

