% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clone.environment}
\alias{clone.environment}
\title{Deep copy of an environment}
\usage{
clone.environment(env, use.copied.ref = FALSE, all.names = TRUE)
}
\arguments{
\item{env}{the environment to be cloned}

\item{use.copied.ref}{internal}
}
\description{
Deep copy of an environment
}

