% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorepoint.R
\name{disable.restore.points}
\alias{disable.restore.points}
\title{Globally disable or enable restore points}
\usage{
disable.restore.points(disable = TRUE)
}
\arguments{
\item{disable}{if TRUE globaly disable restore points. This speeds up calls to restore.point quickly. Is faster than set.storing(FALSE), but has no informative messages when restore.point is called from the global env.}
}
\description{
Globally disable or enable restore points
}
