% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/restorepoint.R
\name{get.stored.object.list}
\alias{get.stored.object.list}
\title{Retrieves the list of all restore.points with the stored objects}
\usage{
get.stored.object.list()
}
\description{
Retrieves the list of all restore.points with the stored objects
}

