\name{restore.point.browser}
\alias{restore.point.browser}
\title{Examing a restore point by invoking the browser}
\usage{
  restore.point.browser(name, was.forced = FALSE,
    message.text = paste("restore point", name, ", press ESC to return."),
    deep.copy = get.restore.point.options()$deep.copy)
}
\description{
  Examing a restore point by invoking the browser
}

