% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorepoint.R
\name{get.stored.dots}
\alias{get.stored.dots}
\title{Returns the ellipsis (...) that has been stored in restore.point name as a list}
\usage{
get.stored.dots(name, deep.copy = FALSE)
}
\arguments{
\item{name}{the name whith which restore.point or store.objects has been called.}

\item{deep.copy}{shall a deep copy of stored objects be made}
}
\description{
Returns the ellipsis (...) that has been stored in restore.point name as a list
}

