\name{get.stored.dots}
\alias{get.stored.dots}
\title{Returns the ellipsis (...) that has been stored in restore.point name as a list}
\usage{
  get.stored.dots(name)
}
\arguments{
  \item{name}{the name whith which restore.point or
  store.objects has been called.}
}
\description{
  Returns the ellipsis (...) that has been stored in
  restore.point name as a list
}

