\name{set.storing}
\alias{set.storing}
\title{Set whether objects shall be stored or not}
\usage{
  set.storing(storing = TRUE)
}
\arguments{
  \item{store}{if FALSE don't store objects if
  restore.point or store.objects is called. May save time.
  If TRUE (default) turn on storage again.}
}
\description{
  Set whether objects shall be stored or not
}

